move _scaler function from plot to calc step
