+++ src/traces/image/plot.js
@@ -10,53 +10,8 @@
 
 var d3 = require('d3');
 var Lib = require('../../lib');
-var isNumeric = require('fast-isnumeric');
 var xmlnsNamespaces = require('../../constants/xmlns_namespaces');
 var constants = require('./constants');
-
-function scale(zero, factor, min, max) {
-    return function(c) {
-        c = (c - zero) * factor;
-        c = Lib.constrain(c, min, max);
-        return c;
-    };
-}
-
-function constrain(min, max) {
-    return function(c) { return Lib.constrain(c, min, max);};
-}
-
-// Generate a function to scale color components according to zmin/zmax and the colormodel
-var makeScaler = function(trace) {
-    var colormodel = trace.colormodel;
-    var n = colormodel.length;
-    var cr = constants.colormodel[colormodel];
-
-    var s = [];
-    // Loop over all color components
-    for(var k = 0; k < n; k++) {
-        if(cr.min[k] !== trace.zmin[k] || cr.max[k] !== trace.zmax[k]) {
-            s.push(scale(
-                trace.zmin[k],
-                (cr.max[k] - cr.min[k]) / (trace.zmax[k] - trace.zmin[k]),
-                cr.min[k],
-                cr.max[k]
-            ));
-        } else {
-            s.push(constrain(cr.min[k], cr.max[k]));
-        }
-    }
-
-    return function(pixel) {
-        var c = pixel.slice(0, n);
-        for(var k = 0; k < n; k++) {
-            var ck = c[k];
-            if(!isNumeric(ck)) return false;
-            c[k] = s[k](ck);
-        }
-        return c;
-    };
-};
 
 module.exports = function plot(gd, plotinfo, cdimage, imageLayer) {
     var xa = plotinfo.xaxis;
@@ -136,7 +91,6 @@ module.exports = function plot(gd, plotinfo, cdimage, imageLayer) {
         var ipx = function(i) {return Lib.constrain(Math.round(xa.c2p(x0 + i * dx) - left), 0, imageWidth);};
         var jpx = function(j) {return Lib.constrain(Math.round(ya.c2p(y0 + j * dy) - top), 0, imageHeight);};
 
-        trace._scaler = makeScaler(trace);
         var fmt = constants.colormodel[trace.colormodel].fmt;
         var c;
         for(i = 0; i < cd0.w; i++) {
